unified dataproviders for all datasets and removed classes as method from ABC
