+++ pytorch_learning_tools/data_providers/DataProviderABC.py
@@ -8,15 +8,8 @@ class DataProviderABC(ABC):
     @abstractmethod
     def splits(self):
         pass
-    
-    # how many classes are we predicting?
-    @property
-    @abstractmethod
-    def classes(self):
-        pass
 
-    # get a single data point using unique id
-    # not how you want to iterate over data but useful for inspection
+    # get a single data point using unique id - not how you want to iterate over data but useful for inspection
     @abstractmethod
     def __getitem__(self, unique_id):
         pass
